+++ .env
@@ -26,3 +26,10 @@ APP_SECRET=2e818020e76cfe96967a67051e5a2a2f
 # Configure your db driver and server_version in config/packages/doctrine.yaml
 DATABASE_URL=mysql://simplepay:Mohit_321@127.0.0.1:3306/simplepay
 ###< doctrine/doctrine-bundle ###
+
+###> symfony/messenger ###
+# Choose one of the transports below
+# MESSENGER_TRANSPORT_DSN=amqp://guest:guest@localhost:5672/%2f/messages
+# MESSENGER_TRANSPORT_DSN=doctrine://default
+# MESSENGER_TRANSPORT_DSN=redis://localhost:6379/messages
+###< symfony/messenger ###
