+++ composer.json
@@ -14,6 +14,7 @@
         "symfony/form": "4.3.*",
         "symfony/framework-bundle": "4.3.*",
         "symfony/maker-bundle": "^1.13",
+        "symfony/messenger": "4.3.*",
         "symfony/orm-pack": "^1.0",
         "symfony/phpunit-bridge": "^4.3",
         "symfony/validator": "4.3.*",
