+++ src/Entity/User.php
@@ -3,19 +3,14 @@
 namespace App\Entity;
 
 use Doctrine\ORM\Mapping as ORM;
-use Symfony\Component\Validator\Constraints as Assert;
+Use App\Entity\Common\IndetifiableEntity;
 
 /**
  * @ORM\Entity(repositoryClass="App\Repository\UserRepository")
  */
 class User
 {
-    /**
-     * @ORM\Id()
-     * @ORM\GeneratedValue()
-     * @ORM\Column(type="integer")
-     */
-    private $id;
+    
 
     /**
      * @ORM\Column(type="string", length=255)
@@ -47,21 +42,6 @@ class User
      * @ORM\Column(type="float", nullable=true, options={"default" : 0})
      */
     private $balance;
-
-    /**
-     * @ORM\Column(type="datetime")
-     */
-    private $created_at;
-
-    /**
-     * @ORM\Column(type="datetime")
-     */
-    private $last_modified_at;
-
-    public function getId(): ?int
-    {
-        return $this->id;
-    }
 
     public function getName(): ?string
     {
@@ -131,30 +111,6 @@ class User
     public function setBalance(?float $balance): self
     {
         $this->balance = $balance;
-
-        return $this;
-    }
-
-    public function getCreatedAt(): ?\DateTimeInterface
-    {
-        return $this->created_at;
-    }
-
-    public function setCreatedAt(\DateTimeInterface $created_at): self
-    {
-        $this->created_at = $created_at;
-
-        return $this;
-    }
-
-    public function getLastModifiedAt(): ?\DateTimeInterface
-    {
-        return $this->last_modified_at;
-    }
-
-    public function setLastModifiedAt(\DateTimeInterface $last_modified_at): self
-    {
-        $this->last_modified_at = $last_modified_at;
 
         return $this;
     }
